refactor(api): api 파일 구조 수정;
