+++ src/recoil/cartItems.ts
@@ -1,4 +1,4 @@
-import { getCartItems } from "@/apis/cart";
+import { getCartItems } from "@/auth/apis/cart";
 import { CartItem } from "@/types/cart";
 import { atom, selector } from "recoil";
